update sample to use geo-compose module
